converting arg to ts
